Rename *context files to *state
