+++ runtime-light/server/http/http-server-state.cpp
@@ -2,10 +2,10 @@
 // Copyright (c) 2024 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/stdlib/fork/fork-context.h"
+#include "runtime-light/server/http/http-server-state.h"
 
 #include "runtime-light/state/instance-state.h"
 
-ForkInstanceState &ForkInstanceState::get() noexcept {
-  return InstanceState::get().fork_instance_state;
+HttpServerInstanceState &HttpServerInstanceState::get() noexcept {
+  return InstanceState::get().http_server_instance_state;
 }
