+++ runtime-light/server/server.cmake
@@ -1,2 +1,2 @@
 prepend(RUNTIME_LIGHT_SERVER_SRC server/ http/init-functions.cpp
-        http/http-server-context.cpp job-worker/job-worker-server-context.cpp)
+        http/http-server-state.cpp job-worker/job-worker-server-state.cpp)
