+++ runtime-light/state/init-functions.cpp
@@ -11,7 +11,7 @@
 #include "runtime-light/coroutine/task.h"
 #include "runtime-light/k2-platform/k2-api.h"
 #include "runtime-light/server/init-functions.h"
-#include "runtime-light/server/job-worker/job-worker-server-context.h"
+#include "runtime-light/server/job-worker/job-worker-server-state.h"
 #include "runtime-light/state/instance-state.h"
 #include "runtime-light/streams/streams.h"
 #include "runtime-light/tl/tl-core.h"
