+++ runtime-light/state/instance-state.h
@@ -17,18 +17,18 @@
 #include "runtime-light/coroutine/task.h"
 #include "runtime-light/k2-platform/k2-api.h"
 #include "runtime-light/scheduler/scheduler.h"
-#include "runtime-light/server/http/http-server-context.h"
-#include "runtime-light/server/job-worker/job-worker-server-context.h"
-#include "runtime-light/stdlib/crypto/crypto-context.h"
-#include "runtime-light/stdlib/curl/curl-context.h"
-#include "runtime-light/stdlib/file/file-stream-context.h"
-#include "runtime-light/stdlib/fork/fork-context.h"
-#include "runtime-light/stdlib/job-worker/job-worker-client-context.h"
+#include "runtime-light/server/http/http-server-state.h"
+#include "runtime-light/server/job-worker/job-worker-server-state.h"
+#include "runtime-light/stdlib/crypto/crypto-state.h"
+#include "runtime-light/stdlib/curl/curl-state.h"
+#include "runtime-light/stdlib/file/file-stream-state.h"
+#include "runtime-light/stdlib/fork/fork-state.h"
+#include "runtime-light/stdlib/job-worker/job-worker-client-state.h"
 #include "runtime-light/stdlib/output/output-buffer.h"
-#include "runtime-light/stdlib/regex/regex-context.h"
-#include "runtime-light/stdlib/rpc/rpc-context.h"
-#include "runtime-light/stdlib/string/string-context.h"
-#include "runtime-light/stdlib/system/system-context.h"
+#include "runtime-light/stdlib/regex/regex-state.h"
+#include "runtime-light/stdlib/rpc/rpc-state.h"
+#include "runtime-light/stdlib/string/string-state.h"
+#include "runtime-light/stdlib/system/system-state.h"
 
 inline constexpr uint64_t INVALID_PLATFORM_DESCRIPTOR = 0;
