+++ runtime-light/stdlib/crypto/crypto-state.cpp
@@ -2,7 +2,7 @@
 //  Copyright (c) 2024 LLC «V Kontakte»
 //  Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/stdlib/crypto/crypto-context.h"
+#include "runtime-light/stdlib/crypto/crypto-state.h"
 
 #include "runtime-light/state/instance-state.h"
