+++ runtime-light/stdlib/curl/curl-state.cpp
@@ -2,10 +2,10 @@
 // Copyright (c) 2024 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/stdlib/system/system-context.h"
+#include "runtime-light/stdlib/curl/curl-state.h"
 
 #include "runtime-light/state/instance-state.h"
 
-SystemInstanceState &SystemInstanceState::get() noexcept {
-  return InstanceState::get().system_instance_state;
+CurlInstanceState &CurlInstanceState::get() noexcept {
+  return InstanceState::get().curl_instance_state;
 }
