+++ runtime-light/stdlib/file/file-stream-state.cpp
@@ -2,10 +2,10 @@
 // Copyright (c) 2024 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/stdlib/curl/curl-context.h"
+#include "runtime-light/stdlib/file/file-stream-state.h"
 
 #include "runtime-light/state/instance-state.h"
 
-CurlInstanceState &CurlInstanceState::get() noexcept {
-  return InstanceState::get().curl_instance_state;
+FileStreamInstanceState &FileStreamInstanceState::get() noexcept {
+  return InstanceState::get().file_stream_instance_state;
 }
