+++ runtime-light/stdlib/job-worker/job-worker-api.cpp
@@ -15,11 +15,11 @@
 #include "runtime-light/coroutine/awaitable.h"
 #include "runtime-light/coroutine/task.h"
 #include "runtime-light/k2-platform/k2-api.h"
-#include "runtime-light/server/job-worker/job-worker-server-context.h"
+#include "runtime-light/server/job-worker/job-worker-server-state.h"
 #include "runtime-light/state/instance-state.h"
 #include "runtime-light/stdlib/component/component-api.h"
-#include "runtime-light/stdlib/fork/fork-context.h"
-#include "runtime-light/stdlib/job-worker/job-worker-client-context.h"
+#include "runtime-light/stdlib/fork/fork-state.h"
+#include "runtime-light/stdlib/job-worker/job-worker-client-state.h"
 #include "runtime-light/stdlib/job-worker/job-worker.h"
 #include "runtime-light/streams/streams.h"
 #include "runtime-light/tl/tl-core.h"
