+++ runtime-light/stdlib/job-worker/job-worker-client-state.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2024 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/stdlib/job-worker/job-worker-client-context.h"
+#include "runtime-light/stdlib/job-worker/job-worker-client-state.h"
 
 #include "runtime-light/state/instance-state.h"
