+++ runtime-light/stdlib/system/system-state.cpp
@@ -2,10 +2,10 @@
 // Copyright (c) 2024 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime-light/server/http/http-server-context.h"
+#include "runtime-light/stdlib/system/system-state.h"
 
 #include "runtime-light/state/instance-state.h"
 
-HttpServerInstanceState &HttpServerInstanceState::get() noexcept {
-  return InstanceState::get().http_server_instance_state;
+SystemInstanceState &SystemInstanceState::get() noexcept {
+  return InstanceState::get().system_instance_state;
 }
